EESSI2-146 og EESSI2-160: Opplysninger fra aareg, tps og ereg. Data transfer objects til frontend.

+++ gui-tjenester/src/main/java/no/nav/melosys/tjenester/gui/dto/ArbeidsforholdDto.java
@@ -98,7 +98,7 @@ public class ArbeidsforholdDto {
             permisjonsPeriode.setTom(tilLocalDate(p.getPermisjonsPeriode().getTom()));
             permisjon.setPermisjonsPeriode(permisjonsPeriode);
 
-            permisjon.setPermisjonsprosent(p.getPermisjonsprosent().floatValue());
+            permisjon.setPermisjonsprosent(p.getPermisjonsprosent());
             permisjon.setPermisjonOgPermittering(p.getPermisjonOgPermittering().getValue());
         }
